# guangdong-energy-monitor/docs/API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: 广东省电厂监控平台 API 契约
  description: |
    华电集团广东电厂监控平台后端接口契约文档。
    所有接口统一通过 Gateway（8080）访问，需携带 JWT Token。
    所有响应统一使用 R<T> 包装结构。
    包含认证授权、用户管理、角色管理、组织架构、电厂管理、水文站管理、
    GIS 空间服务、报警管理、预警管理、规则管理、统计分析、预测服务、数据采集共13个控制器、54个REST端点。
  version: 1.0.0
  contact:
    name: 华电集团广东电厂监控平台研发组
servers:
  - url: http://localhost:8080
    description: 本地开发环境（Gateway 入口）
  - url: https://api.gd-energy.example.com
    description: 生产环境
security:
  - BearerAuth: []
...
paths:
...
  /api/alarms:
    get:
      tags: [报警管理]
      summary: 报警列表（分页）
      description: 按级别、状态、电厂、时间范围分页查询报警
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: level
          in: query
          description: 报警级别
          schema:
            type: string
            enum: [emergency, high, medium, low]
        - name: status
          in: query
          description: 处置状态
          schema:
            type: string
            enum: [unconfirmed, confirmed, resolved]
        - name: plantId
          in: query
          description: 电厂ID
          schema:
            type: integer
            format: int64
        - name: startTime
          in: query
          description: 触发开始时间（yyyy-MM-dd HH:mm:ss）
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          description: 触发结束时间（yyyy-MM-dd HH:mm:ss）
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResultAlarm'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/alarms/{id}:
    get:
      tags: [报警管理]
      summary: 报警详情
      description: 根据报警ID查询详情
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Alarm'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/alarms/{id}/confirm:
    put:
      tags: [报警管理]
      summary: 确认报警
      description: 运维人员确认报警，状态由 unconfirmed 变为 confirmed
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: handler
          in: query
          required: true
          description: 确认人
          schema:
            type: string
      responses:
        '200':
          description: 确认成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Alarm'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/alarms/{id}/resolve:
    put:
      tags: [报警管理]
      summary: 处理报警
      description: 运维人员处置报警，状态变为 resolved
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: handler
          in: query
          required: true
          description: 处理人
          schema:
            type: string
        - name: remark
          in: query
          required: false
          description: 处理备注
          schema:
            type: string
      responses:
        '200':
          description: 处理成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Alarm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # 预警管理 (WarningController) - 3 endpoints
  # ============================================================
  /api/warnings:
    get:
      tags: [预警管理]
      summary: 预警列表（分页）
      description: 查询气象、水文、火情等预警信息
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: level
          in: query
          description: 预警级别
          schema:
            $ref: '#/components/schemas/WarningLevel'
        - name: type
          in: query
          description: 预警类型
          schema:
            $ref: '#/components/schemas/WeatherWarningType'
        - name: status
          in: query
          description: 状态
          schema:
            type: string
            enum: [active, expired, cancelled]
        - name: plantId
          in: query
          description: 电厂ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResultWarning'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/warnings/{id}:
    get:
      tags: [预警管理]
      summary: 预警详情
      description: 根据预警ID查询详情
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Warning'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/warnings/{id}/cancel:
    put:
      tags: [预警管理]
      summary: 取消预警
      description: 取消指定预警
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Warning'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # 规则管理 (RuleManageController) - 2 endpoints
  # ============================================================
...
  /api/predict/water-level/{stationId}:
    get:
      tags: [预测服务]
      summary: 水位预测
      description: 基于机器学习模型预测指定水文站未来水位
      parameters:
        - name: stationId
          in: path
          required: true
          description: 水文站ID
          schema:
            type: integer
            format: int64
        - name: forecastHours
          in: query
          required: false
          description: 预测时长（小时），默认24小时
          schema:
            type: integer
            default: 24
        - name: modelType
          in: query
          required: false
          description: 模型类型
          schema:
            type: string
            enum: [lstm, xgboost, ensemble]
            default: ensemble
      responses:
        '200':
          description: 预测成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PredictionResponseDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/predict/power/{plantId}:
    get:
      tags: [预测服务]
      summary: 发电量预测
      description: 基于机器学习模型预测指定电厂未来发电量
      parameters:
        - name: plantId
          in: path
          required: true
          description: 电厂ID
          schema:
            type: integer
            format: int64
        - name: forecastHours
          in: query
          required: false
          description: 预测时长（小时），默认24小时
          schema:
            type: integer
            default: 24
        - name: modelType
          in: query
          required: false
          description: 模型类型
          schema:
            type: string
            enum: [lstm, xgboost, ensemble]
            default: ensemble
      responses:
        '200':
          description: 预测成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PredictionResponseDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # 数据采集 (CollectorController) - 5 endpoints
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer Token 认证，登录后获取 token，请求头携带 Authorization: Bearer <token>"
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码（从1开始）
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 每页条数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdParam:
      name: id
      in: path
      required: true
      description: 主键ID
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: 请求参数错误（400 PARAM_ERROR）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
    Unauthorized:
      description: 未认证或Token无效（401 UNAUTHORIZED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
...
    NotFound:
      description: 资源不存在（404 NOT_FOUND）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
...
  schemas:
...
    R:
      type: object
      description: |
        统一响应包装结构 R<T>。
        错误码：200 SUCCESS, 400 PARAM_ERROR, 401 UNAUTHORIZED,
        402 TOKEN_EXPIRED, 403 FORBIDDEN, 404 NOT_FOUND,
        500 BUSINESS_ERROR, 5000 SYSTEM_ERROR
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 200
        msg:
          type: string
          description: 提示信息
          example: 操作成功
        data:
          description: 业务数据（具体结构见各接口）
          nullable: true

    # --------------------------------------------------------
    # 分页结果 PageResult<T>
    # --------------------------------------------------------
    PageResult:
      type: object
      description: 分页结果通用结构 PageResult<T>
      properties:
        list:
          type: array
          description: 数据列表
          items: {}
        total:
          type: integer
          format: int64
          description: 总记录数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页条数
...
    PageResultAlarm:
      allOf:
        - $ref: '#/components/schemas/PageResult'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/Alarm'
    PageResultWarning:
      allOf:
        - $ref: '#/components/schemas/PageResult'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/Warning'

    # --------------------------------------------------------
    # 认证与用户
    # --------------------------------------------------------
...
    Alarm:
      type: object
      description: 报警信息
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        level:
          type: string
          description: 报警级别
          enum: [emergency, high, medium, low]
        status:
          type: string
          description: 处置状态
          enum: [unconfirmed, confirmed, resolved]
        content:
          type: string
          description: 报警内容
          example: 1号机组主变压器温度超限（85℃）
        plantId:
          type: integer
          format: int64
          description: 关联电厂ID
        plantName:
          type: string
          description: 关联电厂名称
        triggerTime:
          type: string
          format: date-time
          description: 触发时间
        confirmTime:
          type: string
          format: date-time
          description: 确认时间
          nullable: true
        resolveTime:
          type: string
          format: date-time
          description: 处理完成时间
          nullable: true
        handler:
          type: string
          description: 处理人
          nullable: true
        remark:
          type: string
          description: 处理备注
          nullable: true
        createTime:
          type: string
          format: date-time
          description: 创建时间

    # --------------------------------------------------------
    # 预警
    # --------------------------------------------------------
    Warning:
      type: object
      description: 预警信息
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        level:
          $ref: '#/components/schemas/WarningLevel'
        type:
          $ref: '#/components/schemas/WeatherWarningType'
        content:
          type: string
          description: 预警内容
        plantId:
          type: integer
          format: int64
          description: 关联电厂ID
        plantName:
          type: string
          description: 关联电厂名称
        startTime:
          type: string
          format: date-time
          description: 预警开始时间
        endTime:
          type: string
          format: date-time
          description: 预警结束时间
        status:
          type: string
          description: 状态
          enum: [active, expired, cancelled]
        createTime:
          type: string
          format: date-time
          description: 创建时间

    # --------------------------------------------------------
    # 气象数据
    # --------------------------------------------------------
...
    PredictionResponseDTO:
      type: object
      description: 预测服务响应
      properties:
        predictType:
          type: string
          description: 预测类型（water_level / power）
        targetId:
          type: integer
          format: int64
          description: 目标ID（水文站ID或电厂ID）
        forecastTimes:
          type: array
          description: 预测时间序列
          items:
            type: string
            format: date-time
        forecastValues:
          type: array
          description: 预测值序列
          items:
            type: number
            format: double
        upperBound:
          type: array
          description: 置信上界
          items:
            type: number
            format: double
        lowerBound:
          type: array
          description: 置信下界
          items:
            type: number
            format: double
        confidence:
          type: number
          format: double
          description: 置信度（0~1）
        modelUsed:
          type: string
          description: 使用的模型
        fallback:
          type: boolean
          description: 是否使用降级策略
        generatedAt:
          type: string
          format: date-time
          description: 预测生成时间

    # --------------------------------------------------------
    # 时间序列数据
    # --------------------------------------------------------
...
    WarningLevel:
      type: string
      description: 预警级别
      enum:
        - blue
        - yellow
        - orange
        - red
    WeatherWarningType:
      type: string
      description: 预警类型
      enum:
        - typhoon
        - rainstorm
        - high_temp
        - low_temp
        - gale
        - fog
        - haze
        - thunder
        - ice
        - drought
        - flood
        - other
...